1.消息支持按照时间段导出;2.客户端消费数据上报;3.topic支持在线跨集群迁移;4.死信消息支持重消费;5.UI全面支持移动端;6.部署组件支持安全启停;7.增加任务管理模块;8.集群监控支持动态修改，无需重启；等等

+++ mq-cloud/src/main/java/com/sohu/tv/mq/cloud/bo/ConsumerStat.java
@@ -19,7 +19,7 @@ public class ConsumerStat {
 	//单队列最大未处理消息数
 	private String undone1qMsgCount;
 	//未处理消息时延(broker最新消息存储时间-最新消费时间),ms
-	private int undoneDelay;
+	private long undoneDelay;
 	//错误的订阅关系
 	private String sbscription;
 	//记录更新时间
@@ -57,10 +57,10 @@ public class ConsumerStat {
 	public void setUndone1qMsgCount(String undone1qMsgCount) {
 		this.undone1qMsgCount = undone1qMsgCount;
 	}
-	public int getUndoneDelay() {
+	public long getUndoneDelay() {
 		return undoneDelay;
 	}
-	public void setUndoneDelay(int undoneDelay) {
+	public void setUndoneDelay(long undoneDelay) {
 		this.undoneDelay = undoneDelay;
 	}
 	public String getSbscription() {
